+++ app.js
@@ -2,6 +2,8 @@ const path = require('path');
 const express = require('express');
 const bodyParser = require('body-parser');
 
+const errorsController = require('./controllers/errors');
+
 const app = express();
 
 // setting express global configuration value
@@ -15,12 +17,10 @@ const shopRoutes = require('./routes/shop');
 app.use(bodyParser.urlencoded({extended: false}));
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/admin', adminData.routes);
+app.use('/admin', adminData);
 app.use(shopRoutes);
 
-app.use('/', (req, res, next) => {
-    res.status(404).render('404', {pageTitle: 'Page not found'});
-});
+app.use(errorsController.get404);
 
 // visit localhost:3000 in browser
 app.listen(3000);
