+++ routes/admin.js
@@ -1,18 +1,11 @@
-const path = require('path');
 const express = require('express');
+
+const productsController = require('../controllers/products');
 
 const router = express.Router();
 
-const products = [];
+router.get('/add-product', productsController.getAddProduct);
 
-router.get('/add-product', (req, res, next) => {
-    res.render('add-product', {pageTitle: 'Add product', path: '/admin/add-product'});
-});
+router.post('/add-product', productsController.postAddProduct);
 
-router.post('/add-product', (req, res, next) => {
-    products.push({title: req.body.title});
-    res.redirect('/');
-});
-
-exports.routes = router;
-exports.products = products;
+module.exports = router;
