+++ routes/shop.js
@@ -1,12 +1,9 @@
-const path = require('path');
 const express = require('express');
 
-const router = express.Router();
-const adminData = require('./admin');
+const productsController = require('../controllers/products');
 
-router.get('/', (req, res, next) => {
-    const products = adminData.products;
-    res.render('shop', {products: products, pageTitle: 'My products', path: '/'});
-});
+const router = express.Router();
+
+router.get('/', productsController.getProducts);
 
 module.exports = router;
